refactor: modularize music style catalog
